+++ src/pages/Home/Benefits/Path.tsx
@@ -7,7 +7,7 @@ import icon5 from "../../../assets/landing/Icon5_wb.webp";
 import icon6 from "../../../assets/landing/Icon6_wb.webp";
 import icon7 from "../../../assets/landing/Icon7_wb.webp";
 import icon8 from "../../../assets/landing/Icon8_wb.webp";
-import Button from "../../../components/landing/Button";
+import Button from "../common/Button";
 import BenefitsCarousel from "./BenefitsCarousel";
 import Carousel from "./Carousel";
 const Path = () => {
